actualizadas las rutas en archivos diferentes por tipo, organizados los controllers correspondientes en directorio userControllers, refactorizado el código de los userControllers para dividir en controllers más ligeros + models potencialmente reutilizables
